+++ backend/api/models.py
@@ -10,7 +10,7 @@ import json
 from django.utils import timezone
 from django.conf import settings
 from api.services import Providers, ServiceConfig
-from api.tasks import trigger_sync_tasks
+from api.tasks.syncing import trigger_sync_tasks
 from backend.quotas import (
     can_add_account,
     can_add_app,
